# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RideTech API",
        "description": "Документация API для Инноватсияи ояндасоз",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api",
            "description": "API сервер"
        }
    ],
    "paths": {
...
        "/trips": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Получить список поездок пользователя",
                "description": "Возвращает список поездок текущего пользователя с фильтрацией по водителю, пассажиру, статусу и дате.",
                "operationId": "10ba3ec4a1798366bf9e858297b8c635",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Количество элементов на странице",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "driver_id",
                        "in": "query",
                        "description": "Фильтр по ID водителя (только для пассажиров)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "passenger_id",
                        "in": "query",
                        "description": "Фильтр по ID пассажира (только для водителей)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фильтр по статусу поездки",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Фильтр по дате создания поездки (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список поездок",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Trip"
                                            }
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Trips"
                ],
                "summary": "Создать новую поездку",
                "description": "Создаёт новую поездку и возвращает её данные.",
                "operationId": "47dcf5ea0d5a1f23a28fb51e34ad081f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTripRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Поездка успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trip"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ"
                    },
                    "403": {
                        "description": "Доступ запрещён"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/trips/{id}": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Получить информацию о поездке",
                "description": "Возвращает данные конкретной поездки.",
                "operationId": "7c45de8eac6e2b969d441855f39eb437",
                "parameters": [
                    {
                        "name": "trip",
                        "in": "path",
                        "description": "ID поездки",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные поездки",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trip"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ"
                    },
                    "403": {
                        "description": "Доступ запрещён"
                    },
                    "404": {
                        "description": "Поездка не найдена"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Trips"
                ],
                "summary": "Обновить информацию о поездке",
                "description": "Обновляет информацию о поездке, включая статус.",
                "operationId": "6135e941d2681d021e4e80660477ea37",
                "parameters": [
                    {
                        "name": "trip",
                        "in": "path",
                        "description": "ID поездки",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTripRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Поездка успешно обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trip"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ"
                    },
                    "403": {
                        "description": "Доступ запрещён"
                    },
                    "404": {
                        "description": "Поездка не найдена"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Trips"
                ],
                "summary": "Удалить поездку",
                "description": "Удаляет поездку из системы.",
                "operationId": "1fca55452b03229a8d0ae823b0fa5479",
                "parameters": [
                    {
                        "name": "trip",
                        "in": "path",
                        "description": "ID поездки",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Поездка успешно удалена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Trip canceled"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ"
                    },
                    "403": {
                        "description": "Доступ запрещён"
                    },
                    "404": {
                        "description": "Поездка не найдена"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateTripRequest": {
                "title": "Запрос на создание поездки",
                "description": "Данные, необходимые для создания поездки",
                "required": [
                    "pickup_address",
                    "destination_address"
                ],
                "properties": {
                    "pickup_address": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "123 Main St, New York, NY"
                    },
                    "destination_address": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "456 Elm St, Los Angeles, CA"
                    },
                    "preferences": {
                        "type": "string",
                        "example": "No smoking, quiet ride",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateTripRequest": {
                "required": [
                    "status"
                ],
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/CreateTripRequest"
                    },
                    {
                        "properties": {
                            "status": {
                                "description": "Статус поездки",
                                "type": "string",
                                "enum": [
                                    "pending",
                                    "accepted",
                                    "in_progress",
                                    "completed",
                                    "cancelled"
                                ]
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "Trip": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "driver_id": {
                        "type": "integer"
                    },
                    "passenger_id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "BearerAuth": []
        }
    ]
}